% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy}
\alias{proxy}
\title{[Experimental] Compute document/feature proximity}
\usage{
proxy(x, y = NULL, margin = c("documents", "features"),
  method = c("cosine", "correlation", "jaccard", "ejaccard", "dice",
  "edice", "hamman", "simple matching", "faith", "euclidean", "chisquared",
  "hamming", "kullback", "manhattan", "maximum", "canberra", "minkowski"),
  p = 2, min_proxy = NULL, rank = NULL)
}
\arguments{
\item{y}{if a \link{dfm} object is provided, proximity between documents or
features in \code{x} and \code{y} is computed.}

\item{min_proxy}{the minimum proximity value to be recoded.}

\item{rank}{an integer value specifying top-n most proximity values to be
recorded.}
}
\description{
This is an underlying function for \code{textstat_dist} and
\code{textstat_simil} but returns \code{TsparseMatrix}.
}
\keyword{internal}
