% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{simil}
\alias{simil}
\alias{dist}
\title{Compute similiarty/distance between raws or columns of large matrices}
\usage{
simil(x, y = NULL, margin = 1, method = c("cosine", "correlation",
  "jaccard", "ejaccard", "dice", "edice", "hamman", "simple matching",
  "faith"), min_proxy = NULL, rank = NULL)

dist(x, y = NULL, margin = 1, method = c("euclidean", "chisquared",
  "hamming", "kullback", "manhattan", "maximum", "canberra", "minkowski"),
  p = 2)
}
\arguments{
\item{y}{if a \link{matrix} or \link{Matrix} object is provided, proximity
between documents or features in \code{x} and \code{y} is computed.}

\item{min_proxy}{the minimum proximity value to be recoded.}

\item{rank}{an integer value specifying top-n most proximity values to be
recorded.}
}
\description{
Compute similiarty/distance between raws or columns of large matrices
}
\examples{
mt <- Matrix::rsparsematrix(100, 100, 0.01)
simil(mt, method = "cosine")
dist(mt, method = "euclidean")
}
