% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{colZeros}
\alias{colZeros}
\alias{rowZeros}
\title{Count number of zeros in columns and rows of large matrices}
\usage{
colZeros(x)

rowZeros(x)
}
\arguments{
\item{x}{\link{matrix} or \link{Matrix} object}
}
\description{
Produces the same result as applying \code{sum(x == 0)} to each row or column.
}
\examples{
mt <- Matrix::rsparsematrix(100, 100, 0.01)
colZeros(mt)
apply(mt, 2, function(x) sum(x == 0)) # the same
}
