% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Create a pattern matrix for masking}
\usage{
mask(x, y)
}
\arguments{
\item{x, y}{a numeric or character vector to match against each other.}
}
\value{
a sparse logical matrix with \code{TRUE} for matched pairs.
}
\description{
Create a pattern matrix for \code{simil()} or \code{dist()} to enable masking.
}
\examples{
mask(c("a", "b", "c", "d", "e"),
     c("a", "a", "d", "d", "e", "e"))

}
